🚚 Move: leftcomponent, rightcomponent폴더로 정리
