+++ src/components/workdetail/CdCase.jsx
@@ -1,6 +1,6 @@
 import * as C from './CdCase.style';
-import LeftContent from './LeftContent';
-import RightContent from './RightContent';
+import LeftContent from './leftcomponent/LeftContent';
+import RightContent from './rightcommponent/RightContent';
 
 export default function CdCase({ cdCaseImage }) {
   return (
