+++ src/components/workdetail/leftcomponent/Image.jsx
