+++ src/components/workdetail/leftcomponent/Image.style.js
