+++ src/components/workdetail/leftcomponent/LeftContent.jsx
