+++ src/components/workdetail/leftcomponent/LeftContent.style.js
